runs cmd
This replaces list-runs and delete-run.

Delete now moves a run to a hidden directory under the runs dir so
that it can be recovered or purged using 'guild runs purge'.

There's currently no way to recover a run via the CLI - that needs to
be done manually at this time.