Support for template under the table element itself
JSoup is simply awful and Vulcanize reliance on it is an ever-living
bug trap. HTML input can get mangled silently and then buried in a
single huge HTML output. If this proves to be an ongoing source of
bugs, we should rewite it in Python using BeautifulSoup (or some other
parser that doesn't silently modify the DOM!)

Here are the bugs to watch out for:

- Style errors (as selected paths will change)

- JavaScript element select errors

- Other places where template tags (or other "non HTML5 compliant
  tags) are reparented

These will suddenly appear in the "binary" HTML without notice. Not
fun.